- move asset related logic to separate file`assets.ts` - split logging code into separate file `logging.ts` - add `dotnet.wasm` and `dotnet-crypto-worker.js` to mono-config.json - moved `readSymbolMapFile()` call to `startup.ts` - fixed bug in `setup_proxy_console()` which didn't really copy the original `log` and `error` functions of the console and caused recursion on error. - moved `init_crypto` after `mono_wasm_load_config` in `mono_wasm_pre_init_essential_async` - added new resource type `js-module-crypto` and `js-module-threads` - changed the detection of blazor startup sequence to not be disabled by presence of config, but only by config.assets having some assembly in it.
