+++ src/mono/sample/wasm/browser-bench/frame-main.js
@@ -38,11 +38,11 @@ try {
                 console.error(...arguments);
             }
         },
-        onConfigLoaded: () => {
+        onConfigLoaded: (config) => {
             if (window.parent != window) {
                 window.parent.resolveAppStartEvent("onConfigLoaded");
             }
-            // Module.config.diagnosticTracing = true;
+            // config.diagnosticTracing = true;
         },
         onAbort: (error) => {
             wasm_exit(1, error);
