+++ src/mono/wasm/runtime/net6-legacy/exports-legacy.ts
@@ -5,7 +5,7 @@ import { runtimeHelpers } from "../imports";
 import { mono_wasm_load_bytes_into_heap, setB32, setI8, setI16, setI32, setI52, setU52, setI64Big, setU8, setU16, setU32, setF32, setF64, getB32, getI8, getI16, getI32, getI52, getU52, getI64Big, getU8, getU16, getU32, getF32, getF64 } from "../memory";
 import { mono_wasm_new_root_buffer, mono_wasm_new_root, mono_wasm_new_external_root, mono_wasm_release_roots } from "../roots";
 import { mono_run_main, mono_run_main_and_exit } from "../run";
-import { mono_wasm_setenv, mono_wasm_load_data_archive, mono_wasm_load_config, mono_load_runtime_and_bcl_args } from "../startup";
+import { mono_wasm_setenv, mono_wasm_load_config, mono_load_runtime_and_bcl_args } from "../startup";
 import { js_string_to_mono_string, conv_string, js_string_to_mono_string_root, conv_string_root } from "../strings";
 import { MonoConfig, MonoConfigError } from "../types";
 import { mono_array_to_js_array, unbox_mono_obj, unbox_mono_obj_root, mono_array_root_to_js_array } from "./cs-to-js";
@@ -13,6 +13,7 @@ import { js_typed_array_to_array, js_to_mono_obj, js_typed_array_to_array_root, 
 import { mono_bind_static_method, mono_call_assembly_entry_point } from "./method-calls";
 import { mono_wasm_load_runtime } from "../startup";
 import { BINDINGType, MONOType } from "./export-types";
+import { mono_wasm_load_data_archive } from "../assets";
 
 export function export_mono_api(): MONOType {
     return {
